requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django
uWSGI
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
importlib-metadata==6.1.0
Markdown==3.4.2
pytz==2022.7.1
sqlparse==0.4.3
zipp==3.15.0
django-cors-headers==3.14.0
drf-yasg==1.21.5
python-decouple==3.6